m68k: fold scheduler backends into original sources
